dataflow/requirements.txt
pandas>=1.0.3
seaborn>=0.10.0
matplotlib>=3.2.1
apache-beam[gcp]>=2.19.0
python-dotenv>=0.12.0